# docs/system/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_vehicle_document_api_v1_vehicles__vehicle_id__documents__document_type__post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_vehicle_document_api_v1_vehicles__vehicle_id__documents__document_type__post",
        "type": "object"
      },
...
      "FuelType": {
        "description": "Fuel type enumeration",
        "enum": [
          "GASOLINE",
          "DIESEL",
          "HYBRID",
          "ELECTRIC"
        ],
        "title": "FuelType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VehicleCreate": {
        "description": "Schema for creating new vehicles.",
        "properties": {
          "brand": {
            "maxLength": 60,
            "title": "Brand",
            "type": "string"
          },
          "current_mileage": {
            "default": 0,
            "minimum": 0.0,
            "title": "Current Mileage",
            "type": "integer"
          },
          "fuel_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FuelType"
              }
            ],
            "default": "GASOLINE"
          },
          "inspection_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Expiry Date"
          },
          "insurance_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insurance Expiry Date"
          },
          "model": {
            "maxLength": 60,
            "title": "Model",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "registration_number": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Registration Number",
            "type": "string"
          },
          "seating_capacity": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Seating Capacity",
            "type": "integer"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/VehicleStatus"
              }
            ],
            "default": "ACTIVE"
          },
          "tax_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Expiry Date"
          },
          "vehicle_type": {
            "$ref": "#/components/schemas/VehicleType"
          },
          "year_manufactured": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Manufactured"
          }
        },
        "required": [
          "vehicle_type",
          "brand",
          "model",
          "seating_capacity",
          "registration_number"
        ],
        "title": "VehicleCreate",
        "type": "object"
      },
      "VehicleDocumentExpiryNotification": {
        "description": "Notification payload for expiring vehicle documents.",
        "properties": {
          "days_until_expiry": {
            "title": "Days Until Expiry",
            "type": "integer"
          },
          "document_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Path"
          },
          "document_type": {
            "$ref": "#/components/schemas/VehicleDocumentType"
          },
          "document_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Url"
          },
          "expiry_date": {
            "format": "date",
            "title": "Expiry Date",
            "type": "string"
          },
          "registration_number": {
            "title": "Registration Number",
            "type": "string"
          },
          "vehicle_id": {
            "title": "Vehicle Id",
            "type": "integer"
          }
        },
        "required": [
          "vehicle_id",
          "registration_number",
          "document_type",
          "expiry_date",
          "days_until_expiry"
        ],
        "title": "VehicleDocumentExpiryNotification",
        "type": "object"
      },
      "VehicleDocumentType": {
        "description": "Vehicle document categories supported by the system.",
        "enum": [
          "tax",
          "insurance",
          "inspection"
        ],
        "title": "VehicleDocumentType",
        "type": "string"
      },
      "VehicleDocumentUploadResponse": {
        "description": "Response returned after successfully uploading a vehicle document.",
        "properties": {
          "document_path": {
            "title": "Document Path",
            "type": "string"
          },
          "document_type": {
            "$ref": "#/components/schemas/VehicleDocumentType"
          },
          "document_url": {
            "title": "Document Url",
            "type": "string"
          },
          "vehicle_id": {
            "title": "Vehicle Id",
            "type": "integer"
          }
        },
        "required": [
          "vehicle_id",
          "document_type",
          "document_path",
          "document_url"
        ],
        "title": "VehicleDocumentUploadResponse",
        "type": "object"
      },
...
      "VehicleRead": {
        "description": "Schema for returning vehicle data via the API.",
        "properties": {
          "brand": {
            "maxLength": 60,
            "title": "Brand",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_mileage": {
            "default": 0,
            "minimum": 0.0,
            "title": "Current Mileage",
            "type": "integer"
          },
          "fuel_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FuelType"
              }
            ],
            "default": "GASOLINE"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "inspection_document_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Document Path"
          },
          "inspection_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Expiry Date"
          },
          "insurance_document_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insurance Document Path"
          },
          "insurance_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insurance Expiry Date"
          },
          "model": {
            "maxLength": 60,
            "title": "Model",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "registration_number": {
            "title": "Registration Number",
            "type": "string"
          },
          "seating_capacity": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Seating Capacity",
            "type": "integer"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/VehicleStatus"
              }
            ],
            "default": "ACTIVE"
          },
          "tax_document_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Document Path"
          },
          "tax_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Expiry Date"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "vehicle_type": {
            "$ref": "#/components/schemas/VehicleType"
          },
          "year_manufactured": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Manufactured"
          }
        },
        "required": [
          "vehicle_type",
          "brand",
          "model",
          "seating_capacity",
          "id",
          "registration_number",
          "created_at",
          "updated_at"
        ],
        "title": "VehicleRead",
        "type": "object"
      },
      "VehicleStatus": {
        "description": "Vehicle status enumeration",
        "enum": [
          "ACTIVE",
          "MAINTENANCE",
          "INACTIVE"
        ],
        "title": "VehicleStatus",
        "type": "string"
      },
      "VehicleStatusUpdate": {
        "description": "Schema for updating only the vehicle status.",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/VehicleStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "VehicleStatusUpdate",
        "type": "object"
      },
      "VehicleType": {
        "description": "Vehicle type enumeration",
        "enum": [
          "SEDAN",
          "VAN",
          "PICKUP",
          "BUS",
          "OTHER"
        ],
        "title": "VehicleType",
        "type": "string"
      },
      "VehicleUpdate": {
        "description": "Schema for updating vehicle details.",
        "properties": {
          "brand": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "current_mileage": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Mileage"
          },
          "fuel_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FuelType"
              },
              {
                "type": "null"
              }
            ]
          },
          "inspection_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Expiry Date"
          },
          "insurance_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insurance Expiry Date"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "registration_number": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Number"
          },
          "seating_capacity": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seating Capacity"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VehicleStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "tax_expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Expiry Date"
          },
          "vehicle_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VehicleType"
              },
              {
                "type": "null"
              }
            ]
          },
          "year_manufactured": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Manufactured"
          }
        },
        "title": "VehicleUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\u0e23\u0e30\u0e1a\u0e1a\u0e08\u0e2d\u0e07\u0e23\u0e16\u0e2a\u0e33\u0e19\u0e31\u0e01\u0e07\u0e32\u0e19 - Vehicle booking and management system API",
    "title": "Office Vehicle Booking System API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/vehicles/": {
      "get": {
        "description": "List vehicles with optional filtering and pagination.",
        "operationId": "list_vehicles_endpoint_api_v1_vehicles__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VehicleStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "vehicle_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VehicleType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Type"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VehicleRead"
                  },
                  "title": "Response List Vehicles Endpoint Api V1 Vehicles  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Vehicles Endpoint",
        "tags": [
          "vehicles"
        ]
      },
      "post": {
        "description": "Create a new vehicle entry.",
        "operationId": "create_vehicle_endpoint_api_v1_vehicles__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Vehicle Endpoint",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/api/v1/vehicles/document-expiry": {
      "get": {
        "description": "Return notifications for vehicle documents expiring soon.",
        "operationId": "get_vehicle_document_expiry_notifications_api_v1_vehicles_document_expiry_get",
        "parameters": [
          {
            "in": "query",
            "name": "within_days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 0,
              "title": "Within Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VehicleDocumentExpiryNotification"
                  },
                  "title": "Response Get Vehicle Document Expiry Notifications Api V1 Vehicles Document Expiry Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Vehicle Document Expiry Notifications",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{vehicle_id}": {
      "delete": {
        "description": "Delete a vehicle from the fleet.",
        "operationId": "delete_vehicle_endpoint_api_v1_vehicles__vehicle_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Vehicle Endpoint",
        "tags": [
          "vehicles"
        ]
      },
      "get": {
        "description": "Retrieve a vehicle by its identifier.",
        "operationId": "get_vehicle_detail_api_v1_vehicles__vehicle_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Vehicle Detail",
        "tags": [
          "vehicles"
        ]
      },
      "patch": {
        "description": "Update vehicle information.",
        "operationId": "update_vehicle_endpoint_api_v1_vehicles__vehicle_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Vehicle Endpoint",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{vehicle_id}/documents/{document_type}": {
      "post": {
        "description": "Upload and attach a document to a vehicle.",
        "operationId": "upload_vehicle_document_api_v1_vehicles__vehicle_id__documents__document_type__post",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "document_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/VehicleDocumentType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_vehicle_document_api_v1_vehicles__vehicle_id__documents__document_type__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Vehicle Document",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{vehicle_id}/status": {
      "patch": {
        "description": "Update only the status of a vehicle.",
        "operationId": "update_vehicle_status_endpoint_api_v1_vehicles__vehicle_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Vehicle Status Endpoint",
        "tags": [
          "vehicles"
        ]
      }
    },
...
  }
}